refactor: 카드, 게임참가자, 결과로 package 분리
