+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -1,5 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
+import blackjack.domain.participant.Participant;
 import lombok.Getter;
 
 import java.util.stream.Collectors;
