+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,5 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.participant.Participant;
 import lombok.Getter;
 
 import java.util.stream.Collectors;
