+++ src/main/java/blackjack/domain/participant/PlayersFactory.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.participant.Player;
 import blackjack.utils.StringUtils;
 
 import java.util.Collections;
