+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,7 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.enums.Denomination;
-import blackjack.enums.Type;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Denomination;
+import blackjack.domain.card.Type;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
