+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,7 +1,12 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.enums.Denomination;
-import blackjack.enums.Type;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Participant;
+import blackjack.domain.participant.Player;
+import blackjack.domain.card.Denomination;
+import blackjack.domain.card.Type;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
